third_party/WebKit/LayoutTests/crypto/aes-key-algorithm-mismatch-expected.txt

CONSOLE ERROR: key.algorithm does not match that of operation
Call encrypt using the wrong AES key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS key.algorithm.name is "AES-GCM"
PASS error is null
PASS successfullyParsed is true

TEST COMPLETE
